Use PatternFly AI Chatbot's ChatbotHeader and MessageBar components
Combines all of the prompt input logic into it's own component
