+++ locales/en/plugin__lightspeed-console-plugin.json
@@ -1,11 +1,9 @@
 {
   "{{kind}} {{name}} in namespace {{namespace}}": "{{kind}} {{name}} in namespace {{namespace}}",
-  "Alert": "Alert",
   "Always review AI generated content prior to use.": "Always review AI generated content prior to use.",
   "An unexpected error occurred": "An unexpected error occurred",
   "Are you sure you want to erase the current chat conversation and start a new chat? This action cannot be undone.": "Are you sure you want to erase the current chat conversation and start a new chat? This action cannot be undone.",
   "Attach": "Attach",
-  "Attach context": "Attach context",
   "Bad response": "Bad response",
   "Cancel": "Cancel",
   "Cancelled": "Cancelled",
@@ -41,6 +39,8 @@
   "Failed to load logs": "Failed to load logs",
   "Failed to load pods": "Failed to load pods",
   "Failed to load scale target": "Failed to load scale target",
+  "Filtered YAML": "Filtered YAML",
+  "Full YAML file": "Full YAML file",
   "Good response": "Good response",
   "History truncated": "History truncated",
   "If this error persists, please contact an administrator. Error details: {{e}}": "If this error persists, please contact an administrator. Error details: {{e}}",
@@ -61,7 +61,6 @@
   "Not authorized": "Not authorized",
   "OpenShift Lightspeed authentication failed. Contact your system administrator for more information.": "OpenShift Lightspeed authentication failed. Contact your system administrator for more information.",
   "OpenShift Lightspeed chat history": "OpenShift Lightspeed chat history",
-  "OpenShift Lightspeed prompt": "OpenShift Lightspeed prompt",
   "OpenShift Lightspeed uses AI technology to help answer your questions. Do not include personal information or other sensitive information in your input. Interactions may be used to improve Red Hat's products or services.": "OpenShift Lightspeed uses AI technology to help answer your questions. Do not include personal information or other sensitive information in your input. Interactions may be used to improve Red Hat's products or services.",
   "Optional": "Optional",
   "Please retry or contact support if the issue persists.": "Please retry or contact support if the issue persists.",
@@ -72,7 +71,9 @@
   "Revert to original": "Revert to original",
   "Save": "Save",
   "Send a message...": "Send a message...",
+  "Send message": "Send message",
   "Stay": "Stay",
+  "Stop generating": "Stop generating",
   "Submit": "Submit",
   "Thank you for your feedback!": "Thank you for your feedback!",
   "The following output was generated when running <2>{{name}}</2> with arguments <5>{{argsFormatted}}</5>.": "The following output was generated when running <2>{{name}}</2> with arguments <5>{{argsFormatted}}</5>.",
