+++ src/components/popover.css
@@ -22,6 +22,8 @@
 }
 
 .ols-plugin__popover {
+  --pf-t--chatbot--background: none;
+
   background-color: var(--pf-t--global--background--color--primary--default);
   border-radius: var(--popover-border-radius);
   border: var(--pf-t--global--border--width--regular) solid var(--pf-t--global--border--color--default);
@@ -74,14 +76,13 @@
   width: var(--popover-button-size);
 }
 
-/* stylelint-disable-next-line selector-disallowed-list */
-.pf-v6-theme-dark .ols-plugin__popover-button {
+.ols-plugin__popover-container--dark .ols-plugin__popover-button {
   background-color: var(--pf-t--global--dark--background--color--100);
   background-image: url('../assets/logo-dark.svg');
 }
 
 .ols-plugin__popover-clear-chat {
-  margin-right: var(--pf-t--global--spacer--md);
+  margin-left: var(--pf-t--global--spacer--md);
 }
 
 .ols-plugin__popover-close, .ols-plugin__popover-control {
